+++ src/test/java/com/guillermomolina/protos/runtime/ProtosActorRequestTest.java
@@ -52,7 +52,7 @@ final class ProtosActorRequestTest {
                 assertInstanceOf(
                         ProtosObjectValue.class,
                         target.reference().representedDelegationParent(fixture.prelude));
-        assertEquals(Set.of("send", "request"), actorRefPrototype.localSlotsSnapshot().keySet());
+        assertEquals(Set.of("send", "request", "stop", "termination"), actorRefPrototype.localSlotsSnapshot().keySet());
         assertTrue(actorRefPrototype.isFrozen());
     }